Refactor webhook processor
Extracted webhook processing to its own file
Return 400 error if webhook hmac is not valid
